разбил код на модули. Создал пустые модули на будущее
