+++ js/utils.js
@@ -1,0 +1,38 @@
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+
+  return Math.floor(result);
+};
+
+const createRandomIdFromRangeGenerator = (min, max) => {
+  const previousValues = [];
+
+  return function () {
+    let currentValue = getRandomInteger(min, max);
+
+    if (previousValues.length >= max - min + 1) {
+      // eslint-disable-next-line no-console
+      console.error(`Перебраны все числа из диапазона от ${min} до ${max}`);
+      return null;
+    }
+
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+
+    previousValues.push(currentValue);
+
+    return currentValue;
+  };
+};
+
+const getRandomArrayElement = (elements) =>
+  elements[getRandomInteger(0, elements.length - 1)];
+
+export {
+  createRandomIdFromRangeGenerator,
+  getRandomArrayElement,
+  getRandomInteger,
+};
